extraida funcion a modulo consultarREST.ts
