+++ restapi/rutas/productoRutas.ts
@@ -1,8 +1,8 @@
 import express, { Request, Response } from 'express';
 import { Producto } from '../modelos/productoModelo';
-import { RequestInfo, RequestInit } from 'node-fetch';
+
 import { Foto } from '../modelos/fotoModelo';
-import fetch from 'node-fetch';
+import consultarREST from './consultarREST';
 
 const router = express.Router()
 
@@ -55,31 +55,5 @@ router.post('/products/add', async (req: Request, res: Response) => {
   await product.save()
   return res.status(201).send(product)
 })
-
-async function consultarREST(html: RequestInfo) {
-  try {
-    // 👇️ const response: Response
-    const response = await fetch(html, {
-      method: 'GET',
-      headers: { Accept: 'application/json', },
-    });
-
-    if (!response.ok) {
-      throw new Error(`Error! status: ${response.status}`);
-    }
-
-    const result = (await response.json());
-    //console.log('result is: ', JSON.stringify(result, null, 4));
-    return result;
-  } catch (error) {
-    if (error instanceof Error) {
-      console.log('error message: ', error.message);
-      return error.message;
-    } else {
-      console.log('unexpected error: ', error);
-      return 'An unexpected error occurred';
-    }
-  }
-}
 
 export { router as productoRouter }
